requirements.txt
fastapi==0.115.8
requests==2.32.3
streamlit==1.42.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
pytorch-lightning==2.5.0.post0
torch==2.6.0
torchvision==0.21.0
uvicorn

